requirements.txt
cssmin>=0.2,<0.3
frozen-flask>=0.11,<0.12
htmlmin>=0.1.5,<0.2
inflection>=0.3,<0.4
markdown>=2.6.1,<2.7
slimit>=0.8.1,<0.9
pyyaml>=3.11,<3.12
